// types/packagesType.tsx
import { bookingType } from "./bookingType";
import { specialistType } from "./specialistType";

export type packagesType= {
  _id:string;
  title: string;
  description: string;
  category: string;
  images: string[];
  price: number;
  specialists: specialistType[];
  bookings: bookingType[];
  createdAt:string;
  updatedAt:string;
  __v:number
}